# Avoid array splatting in Types::Definition