@flash-gordon 2016-11-24
Sorry for the delay, I reckon Piotr is busy with his conference talk atm, but it LGTM :) Thanks ❤️